Add support to the stat fops.
A call to fstat(/dev/tap) will return the stat of the vnode associated with
`/dev/tap'. However, as `/dev/tap' is a self-cloning device, we are more
interrested to the stat of the result clone. This is handled by using the
dev_t associated with the cloned tapN interface in the `st_rdev' field of
`struct stat'.

Modified files:
	sys/net/if_tap.c sys/net/if_tap.h