Restore memory detection code based on SW_MEMCONT_REG.
This code actually works, the hardware is however unable to deals with
32MB of memory on bank 0. Valid combinaison are {4|8|16|64|128} as
pointed out by the documentation.

To deal with 32MB of memory on bank 0, keep the actual behaviour: trust
the bootloader. If it is able to provide us a valid value, use it, else
use the value of SW_MEMCONT_REG.

For brocken board with broken bootloader, it is still possible to use
the compile time hard hardcoded value: MEMSIZE.